requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
openai>=1.12.0
httpx==0.25.0
aiofiles==23.2.1
pydantic>=2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0
